Task 4235: Comment system. Change after review

+++ src/main/java/faang/school/postservice/service/CommentService.java
@@ -20,21 +20,22 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final CommentValidator commentValidator;
     private final CommentMapper commentMapper;
-    private final PostService postService;
 
 
     @Transactional
     public CommentDto createComment(Long postId, CommentDto commentDto) {
         commentValidator.validateUserBeforeCreate(commentDto);
-        Post post = postService.getPostById(postId);
         Comment comment = commentMapper.toEntity(commentDto);
+        Post post = new Post();
+        post.setId(postId);
         comment.setPost(post);
         return commentMapper.toDto(commentRepository.save(comment));
     }
 
     @Transactional
     public CommentDto updateComment(Long commentId, CommentDto commentDto) {
-        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new EntityNotFoundException("Comment with id " + commentId + " not found"));
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new EntityNotFoundException("Comment with id " + commentId + " not found"));
         commentValidator.validateBeforeUpdate(comment, commentDto);
         commentMapper.partialUpdate(commentDto, comment);
         return commentMapper.toDto(comment);
@@ -50,7 +51,9 @@ public class CommentService {
 
     @Transactional
     public void deleteComment(Long commentId) {
-        commentRepository.findById(commentId).orElseThrow(() -> new EntityNotFoundException("Comment with id " + commentId + " not found"));
+        if (!commentRepository.existsById(commentId)) {
+            throw new EntityNotFoundException("Comment with id " + commentId + " not found");
+        }
         commentRepository.deleteById(commentId);
     }
 }
